refactor(map): simplify region rendering seam
